app/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
tensorboard>=2.14.0
gradio==3.50.2
huggingface_hub==0.19.4
Pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
streamlit>=1.20.0
